Rename group fetcher to user info fetcher
