+++ rust/operator-binary/src/user_info_fetcher.rs
